requirements.txt
fastapi==0.116.1
uvicorn[standard]
motor==3.7.1
openai==1.99.1
python-dotenv==1.0.1
spacy
httpx==0.28.1
pydantic==2.11.7
apscheduler==3.11.0
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1.tar.gz
